Moved bucketclient implementation to pkg/ruler/rulestore/bucketclient
