+++ pkg/ruler/rulestore/bucketclient/bucket_client_test.go
@@ -1,4 +1,4 @@
-package rulestore
+package bucketclient
 
 import (
 	"context"
